requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
rarfile==4.1
pdf2docx==0.5.8
docx2pdf==0.1.8
Pillow==10.2.0
